CCMSG-1349: Retries hang forever (#575)

+++ src/test/java/io/confluent/connect/elasticsearch/integration/ElasticsearchConnectorBaseIT.java
@@ -29,6 +29,7 @@ import org.junit.AfterClass;
 import org.junit.Before;
 
 import java.io.IOException;
+import java.net.ConnectException;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
@@ -86,9 +87,15 @@ public class ElasticsearchConnectorBaseIT extends BaseConnectorIT {
   public void cleanup() throws IOException {
     stopConnect();
 
-    if (helperClient != null) {
-      helperClient.deleteIndex(TOPIC);
-      helperClient.close();
+    if (container.isRunning()) {
+      if (helperClient != null) {
+        try {
+          helperClient.deleteIndex(TOPIC);
+          helperClient.close();
+        } catch (ConnectException e) {
+          // Server is already down. No need to close
+        }
+      }
     }
   }
